backend/requirements.txt
Flask
python-dotenv
google-generativeai
Flask-Cors